PlantAppFlask/requirements.txt
flask
flask-cors
tensorflow
Pillow
numpy
opencv-python
